tesser/requirements.txt
pdf2image
pytesseract
Pillow
Flask
python-docx
opencv-python-headless
